chore: answer review round 1 — scope the TypeAdapter rule, split the deriver test module
The flat "never TypeAdapter" agent rule becomes a scoped one at its source
(python_standards.md, regenerated into AGENTS.md): don't design shapes that
need one, but validating into a top-level union owned by an external contract
is its sanctioned use — which is exactly what the input-form deriver's
list-item conversion does, so that adapter stays. TestDerivedListItemsAreNameless
moves to its own module per the one-test-class-per-module rule, with its
.test_durations entries following.
