+++ .test_durations
@@ -7964,8 +7964,8 @@
     "tests/unit/pipelex/pipeline/test_hint_warnings.py::TestHintLintSilence::test_applicable_hints_produce_no_warning": 0.0006,
     "tests/unit/pipelex/pipeline/test_hint_warnings.py::TestHintLintSilence::test_hint_free_crate_produces_no_warning": 0.0004,
     "tests/unit/pipelex/pipeline/test_hint_warnings.py::TestHintLintSilence::test_warned_content_stays_in_the_crate": 0.0007,
-    "tests/unit/pipelex/pipeline/test_input_form_deriver.py::TestDerivedListItemsAreNameless::test_a_concept_item_is_the_nameless_model_and_keeps_its_payload": 0.4948,
-    "tests/unit/pipelex/pipeline/test_input_form_deriver.py::TestDerivedListItemsAreNameless::test_a_scalar_item_is_the_nameless_model": 0.4989,
+    "tests/unit/pipelex/pipeline/test_input_form_list_items.py::TestDerivedListItemsAreNameless::test_a_concept_item_is_the_nameless_model_and_keeps_its_payload": 0.4948,
+    "tests/unit/pipelex/pipeline/test_input_form_list_items.py::TestDerivedListItemsAreNameless::test_a_scalar_item_is_the_nameless_model": 0.4989,
     "tests/unit/pipelex/pipeline/test_input_form_deriver.py::TestInputFormDeriverEscapeHatches::test_chain_ending_at_a_base_absent_from_the_crate_is_unknown": 0.2013,
     "tests/unit/pipelex/pipeline/test_input_form_deriver.py::TestInputFormDeriverEscapeHatches::test_concept_absent_from_the_crate_is_unknown_without_a_class": 0.0187,
     "tests/unit/pipelex/pipeline/test_input_form_deriver.py::TestInputFormDeriverEscapeHatches::test_cycle_guard_emits_unknown_on_revisit": 0.1974,
