+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - **The input-form descriptor and the pipe I/O contracts are typed by the standard now (Breaking for importers)**: The wire models of both artifacts are declared once, by the MTHDS standard's Python client, and Pipelex imports them instead of restating them. `pipelex/pipeline/pipe_io_contracts.py` re-exports `PipeInputContract`, `PipeOutputContract`, `PipeIOContract` and `IOMultiplicity` from `mthds.protocol.pipe_io_contracts`; `pipelex/pipeline/input_form.py` re-exports `FieldKind`, `PipeInputFormDescriptor` and the field models from `mthds.protocol.input_form`; and `pipelex/core/pipes/variable_multiplicity.py` re-exports `PresenceMarker` from the contracts module, since the marker is part of the contract's wire vocabulary. **The wire is unchanged** — every field name, every value, and every absent-versus-null rule is the one the engine already emitted, and the derivation is the same derivation. What moved is which package declares the shape. Importers are affected in three ways: a field descriptor is now one model per kind (`TextField`, `ProseField`, `DateField`, `NumberField`, `BooleanField`, `EnumField`, `DocumentField`, `ImageField`, `ObjectField`, `ListField`, `UnknownField`) under an `InputFormField` union discriminated on `kind`, so a node is narrowed with `isinstance` or a `match` on its class rather than read off a flat model; `FieldKind.is_list` and `FieldKind.is_object` are gone with that narrowing; and the `date` kind's Python attribute is now `datetime`, the name it always had on the wire, rather than `datetime_flag`. Because the engine constructs the standard's own closed shapes, a drift between what it emits and what the standard defines now fails at derivation instead of silently on the wire. See [Input-form descriptor](under-the-hood/input-form-descriptor.md) and [Pipe I/O contracts](under-the-hood/pipe-io-contracts.md).
 - **`PresenceMarker`'s symbol grammar moved off the enum (Breaking)**: The marker enum is the standard's, so the three helpers that are this engine's `.mthds` suffix grammar rather than wire vocabulary became module-level functions in `pipelex/core/pipes/variable_multiplicity.py`: `PresenceMarker.from_symbol(...)` is `presence_from_symbol(symbol=...)`, `marker.symbol` is `presence_symbol(presence=marker)`, and `marker.is_force` is `is_force_presence(presence=marker)`, all keyword-only per this repo's convention. `is_optional` and `is_plain` stay properties on the enum, which the standard declares.
 - **The input-form descriptor's item layer is a type of its own (Breaking for importers)**: The standard now splits the field union in two instead of carrying `name` as an optional slot on one shared base, and `pipelex/pipeline/input_form.py` re-exports both layers. `InputFormField` stays the named union, whose per-kind models (`TextField`, `ObjectField`, …) now require `name: str` rather than admitting `None`; the new `InputFormItem` union is their nameless counterpart (`TextItem`, `ObjectItem`, …), which is what a list's `item` holds and the only shape that position accepts. Two re-exported base classes were renamed with it: `InputFormFieldBase` is now `InputFormItemBase`, and `TextValuedFieldBase` is now `TextValuedItemBase`. **The wire is unchanged** — this engine already stopped emitting `name` on an item; what moved is that the rule is now the shape rather than a convention, so an item built with a name fails at construction and a consumer reading `field.name` gets a `str` with nothing to narrow.
+- **Agent-rule refinement on `TypeAdapter`** (`pipelex/kit/agent_rules/python_standards.md`): the flat ban becomes a scoped one — don't design shapes that need a `TypeAdapter`, but validating into a top-level union or container owned by an external contract (a standard's wire model, a log stream of one-of-many event types) is its sanctioned use, as a module-level singleton. The input-form deriver's list-item conversion is exactly that case and keeps its adapter.
 - **Minimum `mthds` version**: The floor moves to `mthds>=0.11.0`, the releases carrying that split and the descriptor's coherence rules — a top-level field whose `required` contradicts its presence marker, one pairing `gating: true` with an optional marker, and a wire node spelling absence as an explicit `null` are now rejected at the parse, which gates this engine's emission at derivation time rather than on the wire. Everyone downstream of Pipelex inherits that floor.
 
 ## [v0.53.0] - 2026-08-25
