+++ pipelex/kit/agent_rules/python_standards.md
@@ -104,7 +104,7 @@ The deciding question is **does this value ever get serialized** — sent across
 - **frozen `BaseModel`** (`model_config = ConfigDict(frozen=True)`) — immutable value objects that need the full model API (`model_dump`, validators, discriminated unions), must sit in a uniformly-`BaseModel` layer, or must stay type-distinct from a `tuple` in a union.
 - **mutable `BaseModel`** — objects genuinely mutated in place that also want the model API or layer consistency.
 
-Never reach for `TypeAdapter` — needing it means the shape was chosen wrong. And don't convert an existing serialized `BaseModel` to a dataclass just because dataclasses are serialization-safe too: a model that relies on discriminated unions or validators must stay a `BaseModel`. When a serialized dataclass/model field references a type from an optional dependency, alias it to `Any` at runtime so it validates as `Any` without importing the extra (e.g. `if TYPE_CHECKING: from temporalio.common import RetryPolicy` / `else: RetryPolicy = Any`) — a bare forward ref instead raises `PydanticUserError` on construction.
+Don't reach for `TypeAdapter` on a shape you designed — needing it there means the top-level shape was chosen wrong: wrap it in a named `BaseModel`, which brings `model_validate` and the rest for free. Its one legitimate use is validating into a top-level type that is not a model class and not yours to wrap — a discriminated union or container owned by an external contract (a standard's wire model, a log stream of one-of-many event types); build that adapter once at module level and reuse it, and never use a `TypeAdapter` as a runtime substitute for `cast` on a value that is already typed. And don't convert an existing serialized `BaseModel` to a dataclass just because dataclasses are serialization-safe too: a model that relies on discriminated unions or validators must stay a `BaseModel`. When a serialized dataclass/model field references a type from an optional dependency, alias it to `Any` at runtime so it validates as `Any` without importing the extra (e.g. `if TYPE_CHECKING: from temporalio.common import RetryPolicy` / `else: RetryPolicy = Any`) — a bare forward ref instead raises `PydanticUserError` on construction.
 
 Once you've landed on a `BaseModel` (the default for anything that serializes), respect Pydantic v2 standards:
